requirements.txt
aiogram>=3.0
aiosqlite
apscheduler
python-dotenv